src/templates/layered/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
pydantic-settings
sqlalchemy ; extra == "db"
alembic ; extra == "db"
